requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.3
cachetools==4.2.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.3.2
deep-translator==1.11.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==2.2.2
